+++ tests/factory.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 import typing
 import unittest.mock
 
@@ -8,7 +10,9 @@ from matplotlib.figure import Figure
 from pyoma2.algorithms import BaseAlgorithm
 from pyoma2.algorithms.data.result import BaseResult
 from pyoma2.algorithms.data.run_params import BaseRunParams
-from pyoma2.support.geometry import Geometry1
+
+if typing.TYPE_CHECKING:
+    pass
 
 FakeFigure = unittest.mock.MagicMock(spec=Figure)
 FakeAxes = unittest.mock.MagicMock(spec=Axes)
@@ -25,6 +29,16 @@ class FakeResult(BaseResult):
     """FakeResult is a subclass of BaseResult."""
 
     Fn: npt.ArrayLike = np.array([1.0, 2.0, 3.0])
+    Phi: npt.ArrayLike = np.array(
+        [
+            [1.0, 2.0, 3.0],
+            [4.0, 5.0, 6.0],
+            [7.0, 8.0, 9.0],
+            [10.0, 11.0, 12.0],
+            [13.0, 14.0, 15.0],
+            [16.0, 17.0, 18.0],
+        ]
+    )
     result1: int = 1
     result2: str = "test"
 
@@ -43,18 +57,6 @@ class FakeAlgorithm(BaseAlgorithm[FakeRunParams, FakeResult, typing.Iterable[flo
 
     def mpe_from_plot(self, *args, **kwargs) -> typing.Any:
         return np.array([1.0, 2.0, 3.0])
-
-    def plot_mode_g1(
-        self,
-        geo1: Geometry1,
-        mode_numb: int,
-        scaleF: int = 1,
-        view: typing.Literal["3D", "xy", "xz", "yz", "x", "y", "z"] = "3D",
-        remove_fill: bool = True,
-        remove_grid: bool = True,
-        remove_axis: bool = True,
-    ) -> typing.Any:
-        return FakeFigure, FakeAxes
 
 
 class FakeAlgorithm2(FakeAlgorithm):